Move statushelper to an internal package
This is currently only being used by exporterhelper. We can move
it later if it needs to be shared between other packages.
